code/backend/requirements.txt
Django
djangorestframework
drf-yasg
django-cors-headers
mysqlclient